px_hand_pose/iinet/requirements.txt
antialiased_cnns==0.3
h5py==3.11.0
kornia==0.7.3
matplotlib==3.7.5
numpy==1.24.4
onnx==1.17.0
opencv-contrib-python==4.11.0.86
pillow==10.4.0
progressbar33==2.4
pycuda==2025.1.1
pypng==0.20220715.0
PyYAML==6.0.2
scipy==1.10.1
tensorboardX==2.6.2.2

--extra-index-url https://pypi.nvidia.com
tensorrt-cu11==10.12.0.36

torch==2.4.1
torchvision==0.19.1
timm==0.9.12
trimesh==4.5.3
